dataset/python-mutated/graph_renderer.py
"""

"""
from __future__ import annotations
import logging
log = logging.getLogger(__name__)
from ...core.properties import Instance, InstanceDefault
from ...core.validation import error
from ...core.validation.errors import MALFORMED_GRAPH_SOURCE
from ..glyphs import MultiLine, Scatter
from ..graphs import GraphHitTestPolicy, LayoutProvider, NodesOnly
from ..sources import ColumnDataSource
from .glyph_renderer import GlyphRenderer
from .renderer import DataRenderer
__all__ = ('GraphRenderer',)
_DEFAULT_NODE_RENDERER = lambda : GlyphRenderer(glyph=Scatter(), data_source=ColumnDataSource(data=dict(index=[])))
_DEFAULT_EDGE_RENDERER = lambda : GlyphRenderer(glyph=MultiLine(), data_source=ColumnDataSource(data=dict(start=[], end=[])))

class GraphRenderer(DataRenderer):
    """

    """

    def __init__(self, *args, **kwargs) -> None:
        if False:
            return 10
        super().__init__(*args, **kwargs)

    @error(MALFORMED_GRAPH_SOURCE)
    def _check_malformed_graph_source(self):
        if False:
            for i in range(10):
                print('nop')
        missing = []
        if 'index' not in self.node_renderer.data_source.column_names:
            missing.append("Column 'index' is missing in GraphSource.node_renderer.data_source")
        if 'start' not in self.edge_renderer.data_source.column_names:
            missing.append("Column 'start' is missing in GraphSource.edge_renderer.data_source")
        if 'end' not in self.edge_renderer.data_source.column_names:
            missing.append("Column 'end' is missing in GraphSource.edge_renderer.data_source")
        if missing:
            return ' ,'.join(missing) + ' [%s]' % self
    layout_provider = Instance(LayoutProvider, help='\n    An instance of a ``LayoutProvider`` that supplies the layout of the network\n    graph in cartesian space.\n    ')
    node_renderer = Instance(GlyphRenderer, default=_DEFAULT_NODE_RENDERER, help='\n    Instance of a ``GlyphRenderer`` containing an ``XYGlyph`` that will be rendered\n    as the graph nodes.\n    ')
    edge_renderer = Instance(GlyphRenderer, default=_DEFAULT_EDGE_RENDERER, help='\n    Instance of a ``GlyphRenderer`` containing an ``MultiLine`` Glyph that will be\n    rendered as the graph edges.\n    ')
    selection_policy = Instance(GraphHitTestPolicy, default=InstanceDefault(NodesOnly), help='\n    An instance of a ``GraphHitTestPolicy`` that provides the logic for selection\n    of graph components.\n    ')
    inspection_policy = Instance(GraphHitTestPolicy, default=InstanceDefault(NodesOnly), help='\n    An instance of a ``GraphHitTestPolicy`` that provides the logic for inspection\n    of graph components.\n    ')